Enhancement: Extract TraversableContainsTest
